requirements.txt
Babel==2.1.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-phonenumber-field==1.0.0
django-registration==2.0.3
django-toolbelt==0.0.1
gunicorn==19.4.1
phonenumbers==7.1.1
psycopg2==2.6.1
pytz==2015.7
static3==0.6.1
wheel==0.24.0
